+++ Config/Config.php
@@ -29,6 +29,14 @@ class Config
     }
 
     /**
+     * @return bool
+     */
+    public function enabled(): bool
+    {
+        return (bool)$this->scopeConfig->getValue('system/yireo_webp/enabled');
+    }
+
+    /**
      * @return int
      */
     public function getQualityLevel(): int
@@ -50,6 +58,6 @@ class Config
      */
     public function isDebugging(): bool
     {
-        return true;
+        return (bool)$this->scopeConfig->getValue('system/yireo_webp/debug');
     }
 }